+++ src/api/api_routes.py
@@ -2,66 +2,18 @@
 # (will only be on with --api starting option)
 # Currently no API stability guarantees are provided - API may break on any new commit.
 
-import numpy as np
 from fastapi import FastAPI, Body
 from fastapi.exceptions import HTTPException
-from PIL import Image
 from itertools import tee
-import json
 
 import gradio as gr
 
-from modules.api.models import List, Dict
-from modules.api import api
+from typing import Dict, List
 
 from src.common_constants import GenerationOptions as go
-from src.core import core_generation_funnel, CoreGenerationFunnelInp
-from src import backbone
 from src.misc import SCRIPT_VERSION
-from src.api.api_constants import api_defaults, api_forced, api_options, models_to_index
-
-def encode_to_base64(image):
-    if type(image) is str:
-        return image
-    elif type(image) is Image.Image:
-        return api.encode_pil_to_base64(image)
-    elif type(image) is np.ndarray:
-        return encode_np_to_base64(image)
-    else:
-        return ""
-
-def encode_np_to_base64(image):
-    pil = Image.fromarray(image)
-    return api.encode_pil_to_base64(pil)
-
-def to_base64_PIL(encoding: str):
-    return Image.fromarray(np.array(api.decode_base64_to_image(encoding)).astype('uint8'))
-
-
-def api_gen(input_images, client_options):
-
-    default_options = CoreGenerationFunnelInp(api_defaults).values
-
-    #TODO try-catch type errors here
-    for key, value in client_options.items():
-        if key == "model_type":
-            default_options[key] = models_to_index(value)
-            continue
-        default_options[key] = value
-
-    for key, value in api_forced.items():
-        default_options[key.lower()] = value
-        
-    print(f"Processing {str(len(input_images))} images through the API")
-
-    print(default_options)
-
-    pil_images = []
-    for input_image in input_images:
-        pil_images.append(to_base64_PIL(input_image))
-    outpath = backbone.get_outpath()
-    gen_obj = core_generation_funnel(outpath, pil_images, None, None, default_options)
-    return gen_obj
+from src.api.api_constants import api_options, models_to_index
+from api.api_core import api_gen, encode_to_base64
 
 def depth_api(_: gr.Blocks, app: FastAPI):
     @app.get("/depth/version")
@@ -72,7 +24,8 @@ def depth_api(_: gr.Blocks, app: FastAPI):
     async def get_options():
         return {
             "gen_options": [x.name.lower() for x in go],
-            "api_options": api_options
+            "api_options": api_options,
+            "model_names": models_to_index.keys()
         }
 
     @app.post("/depth/generate")
